admin일때  피자 등록,수정,삭제 할수 있도록 수정

+++ pages/api/products/[id].js
@@ -3,9 +3,11 @@ import Product from "../../../models/Product";
 
 export default async function handler(req, res) {
   const {
+    cookies,
     method,
     query: { id },
   } = req;
+  const token = cookies.token;
 
   dbConnect();
   if (method === "GET") {
@@ -16,15 +18,25 @@ export default async function handler(req, res) {
       res.status(500).json(error);
     }
   }
+
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticate!!!");
+    }
     try {
-      const product = await Product.findByIdAndUpdate(id, req.body, {new:true});
+      const product = await Product.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
       res.status(200).json(product);
     } catch (error) {
       res.status(500).json(error);
     }
   }
+
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticate!!!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("피자가 삭제 되었습니다.");
